+++ tests/core/controllers/rolesController.spec.ts
@@ -3,11 +3,11 @@ import { StatusCodes } from 'http-status-codes'
 import mongoose from 'mongoose'
 import request from 'supertest'
 import Application from '../../../src/application'
-import { Item } from '../../../src/core/models'
+import { Role } from '../../../src/core/models'
 import db from '../../testsUtils/db'
-import ItemBuilder from '../../testsUtils/item'
+import RoleBuilder from '../../testsUtils/role'
 
-describe('ItemsController', () => {
+describe('RolesControlles', () => {
   let server: http.Server
 
   beforeAll(async () => {
@@ -28,16 +28,45 @@ describe('ItemsController', () => {
     server.close()
   })
 
+  describe('findAll', () => {
+    it(`Should return 401
+        When uses JWTValidator
+        And call the endpoint GET /roles`, async () => {
+      expect.assertions(1)
+
+      await request(server)
+        .get('/v1/roles')
+        .then((result) => {
+          expect(result.statusCode).toBe(StatusCodes.UNAUTHORIZED)
+        })
+    })
+  })
+
   describe('insert', () => {
     it(`Should return 401
         When uses JWTValidator
-        When call the endpoint POST /items`, async () => {
+        When call the endpoint POST /roles`, async () => {
       expect.assertions(1)
-      const item: Item = new ItemBuilder().build()
+      const role: Role = new RoleBuilder().build()
 
       await request(server)
-        .post('/v1/items')
-        .send(item)
+        .post('/v1/roles')
+        .send(role)
+        .then((result) => {
+          expect(result.statusCode).toBe(StatusCodes.UNAUTHORIZED)
+        })
+    })
+  })
+
+  describe('findById', () => {
+    it(`Should return 401
+        When uses JWTValidator
+        When call the endpoint GET /roles/:id`, async () => {
+      expect.assertions(1)
+      const id: mongoose.Types.ObjectId = new mongoose.Types.ObjectId()
+
+      await request(server)
+        .get('/v1/roles/'.concat(id.toString()))
         .then((result) => {
           expect(result.statusCode).toBe(StatusCodes.UNAUTHORIZED)
         })
@@ -47,14 +76,14 @@ describe('ItemsController', () => {
   describe('update', () => {
     it(`Should return 401
         When uses JWTValidator
-        When call the endpoint PUT /items/:id`, async () => {
+        When call the endpoint PUT /roles/:id`, async () => {
       expect.assertions(1)
       const id: mongoose.Types.ObjectId = new mongoose.Types.ObjectId()
-      const item = new ItemBuilder().name('new name').build()
+      const role = new RoleBuilder().name('new name').build()
 
       await request(server)
-        .put('/v1/items/'.concat(id.toString()))
-        .send(item)
+        .put('/v1/roles/'.concat(id.toString()))
+        .send(role)
         .then((result) => {
           expect(result.statusCode).toBe(StatusCodes.UNAUTHORIZED)
         })
@@ -64,12 +93,12 @@ describe('ItemsController', () => {
   describe('delete', () => {
     it(`Should return 401
         When uses JWTValidator
-        When call the endpoint DELETE /items/:id`, async () => {
+        When call the endpoint DELETE /roles/:id`, async () => {
       expect.assertions(1)
       const id: mongoose.Types.ObjectId = new mongoose.Types.ObjectId()
 
       await request(server)
-        .delete('/v1/items/'.concat(id.toString()))
+        .delete('/v1/roles/'.concat(id.toString()))
         .then((result) => {
           expect(result.statusCode).toBe(StatusCodes.UNAUTHORIZED)
         })