+++ CHANGELOG.md
@@ -8,3 +8,18 @@
 - tests
 - dockerfile and docker-compose
 - github actions CI
+
+## [] version 2.0.0
+
+## Added
+
+- badges for build, test and lint status
+- change configs
+- create JWT validator
+- JWT validator to:
+  - POST /items
+  - PUT /items/:id
+  - DELETE /items/:id
+- create new errors
+- authentication endpoint:
+  - POST /users/authenticate